abstracted out test case helper methods to contest_common
